Review tab - Move getOS function to new file
